Revert "RavenDB-21047 - fixing PR comments#1"
This reverts commit 538925906cddfc9550465079f6a25dccca152442.
